fix: in PDFs, positions are integers

+++ livre-structures/src/point.rs
@@ -4,8 +4,8 @@ use nom::sequence::separated_pair;
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Extract<'_> for Point {
@@ -16,7 +16,7 @@ impl Extract<'_> for Point {
 }
 
 impl Point {
-    pub fn new(x: f32, y: f32) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Self { x, y }
     }
 }
@@ -28,8 +28,8 @@ mod tests {
     use super::*;
 
     #[rstest]
-    #[case(b"0. 0.", Point::new(0.0, 0.0))]
-    #[case(b"-10. 0.", Point::new(-10.0, 0.0))]
+    #[case(b"0 0", Point::new(0, 0))]
+    #[case(b"-10 0", Point::new(-10, 0))]
     fn point(#[case] input: &[u8], #[case] expected: Point) {
         let (_, point) = extract(input).unwrap();
         assert_eq!(expected, point);
